Break apart the main __init__.py file
