+++ src/flickr_url_parser/types.py
